django_API/requirements.txt
asgiref==3.8.1
cssselect==1.2.0
d==0.2.2
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
lxml==5.3.0
Markdown==3.7
Pygments==2.18.0
pyquery==2.0.1
pytz==2024.2
sqlparse==0.5.1
